+++ tablib/formats/_json.py
@@ -3,8 +3,16 @@
 """ Tablib - JSON Support
 """
 
-import simplejson as json
+try:
+	import json  # load system JSON (Python >= 2.6)
+except ImportError:
+	try:
+		import simplejson as json
+	except ImportError:
+		import tablib.packages.simplejson as json # use the vendorized copy
+
 import tablib.core
+
 
 title = 'json'
 extentions = ('json', 'jsn')
@@ -43,5 +51,5 @@ def detect(stream):
 	try:
 		json.loads(stream)
 		return True
-	except json.decoder.JSONDecodeError:
+	except ValueError:
 		return False